[FlexNG] Set MinSizeShouldEncompassIntrinsicSize
SetMinBlockSizeShouldEncompassIntrinsicSize() on a flex item
to allow the item to grow as a result of fragmentation.

In order for this flag to have an effect on flex items inside
a column flex container, we also had to SetIsFixedBlockSize()
to false.

This will be used to expand the flex lines if needed.

Bug: 660611
Change-Id: I8890cc54262c14664706b9cb9efb141a453dceff